+++ src/authz-module/index.scss
@@ -118,7 +118,6 @@
 
 .scope-list {
   max-height: 500px;
-  overflow-y: auto;
 }
 
 .toast-container {
